# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Service Support API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /users:
    get:
      tags:
        - user
      summary: Összes felhasználó lekérdezése
      operationId: getAllUsers
      responses:
        '200':
          description: Sikeres válasz
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Hiányzó vagy hibás token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '403':
          description: Jogosultság hiányzik
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
  /users/register:
    post:
      tags: [ user ]
      summary: Új felhasználó regisztrációja
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '200':
          description: Sikeres regisztráció
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}/ban:
    post:
      tags: [ user ]
      summary: Felhasználó kitiltása
      operationId: banUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sikeres tiltás
  /users/{id}/unban:
    post:
      tags: [ user ]
      summary: Felhasználó tiltásának feloldása
      operationId: unbanUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sikeres feloldás
  /users/{id}/password:
    patch:
      tags: [ user ]
      summary: Jelszó megváltoztatása a régi jelszó megadásával
      operationId: changePassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ oldPassword, newPassword ]
              properties:
                oldPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Sikeres jelszócsere
        '400':
          description: Hibás adatok vagy jelszó nem egyezik
  /users/password-with-token:
    patch:
      tags: [ user ]
      summary: Jelszó megváltoztatása reset tokennel (email alapján)
      operationId: changePasswordWithToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ email, token, newPassword ]
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Sikeres jelszócsere
        '400':
          description: Hibás vagy lejárt token
  /users/{id}/set-password:
    patch:
      tags: [ user ]
      summary: Admin által jelszó beállítása
      operationId: setPassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ password ]
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Jelszó frissítve
  /users/{id}/roles/{roleName}:
    post:
      tags: [ user ]
      summary: Role hozzárendelése felhasználóhoz
      operationId: assignRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sikeres role-hozzárendelés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags: [ user ]
      summary: Role eltávolítása felhasználótól
      operationId: removeRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sikeres role-eltávolítás
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/forgot-password:
    post:
      tags: [ user ]
      summary: Elfelejtett jelszó – token létrehozása
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ email ]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Token generálva vagy meglévő visszaadva
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  validUntil:
                    type: string
                    format: date-time
  /me:
    get:
      tags:
        - user
      summary: Visszaadja a bejelentkezett felhasználó adatait és szerepköreit
      operationId: getCurrentUser
      responses:
        '200':
          description: A bejelentkezett user adatai
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Nincs bejelentkezve vagy érvénytelen token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Nincs bejelentkezve"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterUserRequest:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        password:
          type: string
        position:
          type: string
        roles:
          type: array
          items:
            type: string
...
    User:
      type: object
      required: [ id, email, fullName ]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        fullName:
          type: string
        phoneNumber:
          type: string
          example: "+36123456789"
        position:
          type: string
          example: "Rendszergazda"
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
          example: [ "ROLE_ADMIN", "ROLE_USER" ]
...
    MeResponse:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        fullName:
          type: string
          example: Admin Felhasználó
        roles:
          type: array
          items:
            type: string
          example: [ "ROLE_ADMIN", "ROLE_USER" ]
...